# tests/unit/data/openapi.json
{
  "components": {
    "schemas": {
      "AdvanceRequest": {
        "properties": {
          "choice_id": {
            "title": "Choice Id",
            "type": "string"
          },
          "from_node_id": {
            "title": "From Node Id",
            "type": "string"
          }
        },
        "required": [
          "choice_id",
          "from_node_id"
        ],
        "title": "AdvanceRequest",
        "type": "object"
      },
      "AdvanceResponse": {
        "properties": {
          "image_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Status"
          },
          "new_characters": {
            "items": {
              "$ref": "#/components/schemas/Character"
            },
            "title": "New Characters",
            "type": "array"
          },
          "node": {
            "$ref": "#/components/schemas/NodeDetail"
          }
        },
        "required": [
          "node",
          "new_characters"
        ],
        "title": "AdvanceResponse",
        "type": "object"
      },
...
      "GameDetail": {
        "properties": {
          "art_style": {
            "title": "Art Style",
            "type": "string"
          },
          "characters": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Characters",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_node_id": {
            "title": "Current Node Id",
            "type": "string"
          },
          "endings_reached": {
            "items": {
              "type": "string"
            },
            "title": "Endings Reached",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "nodes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/NodeDetail"
            },
            "title": "Nodes",
            "type": "object"
          },
          "relationships": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Relationships",
            "type": "array"
          },
          "root_node_id": {
            "title": "Root Node Id",
            "type": "string"
          },
          "text_total_input_tokens": {
            "default": 0,
            "title": "Text Total Input Tokens",
            "type": "integer"
          },
          "text_total_output_tokens": {
            "default": 0,
            "title": "Text Total Output Tokens",
            "type": "integer"
          },
          "text_total_requests": {
            "default": 0,
            "title": "Text Total Requests",
            "type": "integer"
          },
          "theme": {
            "additionalProperties": true,
            "title": "Theme",
            "type": "object"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tone": {
            "additionalProperties": true,
            "title": "Tone",
            "type": "object"
          },
          "total_image_cost_usd": {
            "default": 0.0,
            "title": "Total Image Cost Usd",
            "type": "number"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "theme",
          "tone",
          "characters",
          "current_node_id",
          "root_node_id",
          "nodes",
          "endings_reached",
          "art_style",
          "created_at",
          "updated_at"
        ],
        "title": "GameDetail",
        "type": "object"
      },
...
      "GraphEdge": {
        "properties": {
          "child_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Child Id"
          },
          "choice_text": {
            "title": "Choice Text",
            "type": "string"
          },
          "parent_id": {
            "title": "Parent Id",
            "type": "string"
          }
        },
        "required": [
          "parent_id",
          "choice_text"
        ],
        "title": "GraphEdge",
        "type": "object"
      },
      "GraphResponse": {
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Edges",
            "type": "array"
          }
        },
        "required": [
          "edges"
        ],
        "title": "GraphResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JumpRequest": {
        "properties": {
          "target_node_id": {
            "title": "Target Node Id",
            "type": "string"
          }
        },
        "required": [
          "target_node_id"
        ],
        "title": "JumpRequest",
        "type": "object"
      },
...
      "PruneRequest": {
        "properties": {
          "node_id": {
            "title": "Node Id",
            "type": "string"
          }
        },
        "required": [
          "node_id"
        ],
        "title": "PruneRequest",
        "type": "object"
      },
      "RegenerateNodeRequest": {
        "description": "Re-roll the current node by pruning it and re-advancing from the parent.",
        "properties": {},
        "title": "RegenerateNodeRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "par-storygen API"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/games/{game_id}": {
      "delete": {
        "description": "Delete a game save.",
        "operationId": "delete_game_endpoint_api_games__game_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Game Endpoint",
        "tags": [
          "games"
        ]
      },
      "get": {
        "description": "Return full game state.",
        "operationId": "get_game_api_games__game_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Game",
        "tags": [
          "games"
        ]
      }
    },
    "/api/games/{game_id}/advance": {
      "post": {
        "description": "Pick a choice and advance the story.",
        "operationId": "advance_game_api_games__game_id__advance_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Advance Game",
        "tags": [
          "games"
        ]
      }
    },
    "/api/games/{game_id}/endings": {
      "get": {
        "description": "List reached ending node IDs.",
        "operationId": "list_endings_api_games__game_id__endings_get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response List Endings Api Games  Game Id  Endings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Endings",
        "tags": [
          "games"
        ]
      }
    },
    "/api/games/{game_id}/export-book": {
      "post": {
        "description": "Export the current ending path as an HTML book.",
        "operationId": "export_book_endpoint_api_games__game_id__export_book_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Export Book Endpoint Api Games  Game Id  Export Book Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Book Endpoint",
        "tags": [
          "games"
        ]
      }
    },
    "/api/games/{game_id}/export-book/download": {
      "get": {
        "description": "Download the exported HTML book as a file.",
        "operationId": "download_book_api_games__game_id__export_book_download_get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Book",
        "tags": [
          "games"
        ]
      }
    },
    "/api/games/{game_id}/graph": {
      "get": {
        "description": "Return all choice edges for the story graph (visited + unvisited).",
        "operationId": "get_graph_api_games__game_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Graph",
        "tags": [
          "games"
        ]
      }
    },
    "/api/games/{game_id}/jump": {
      "post": {
        "description": "Set the current node to a target node.",
        "operationId": "jump_to_node_api_games__game_id__jump_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JumpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Jump To Node",
        "tags": [
          "games"
        ]
      }
    },
    "/api/games/{game_id}/path": {
      "get": {
        "description": "Return the path from root to target node.",
        "operationId": "get_path_api_games__game_id__path_get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_node_id",
            "required": true,
            "schema": {
              "title": "Target Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NodeDetail"
                  },
                  "title": "Response Get Path Api Games  Game Id  Path Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Path",
        "tags": [
          "games"
        ]
      }
    },
    "/api/games/{game_id}/prune": {
      "post": {
        "description": "Prune a subtree starting at the given node.",
        "operationId": "prune_subtree_endpoint_api_games__game_id__prune_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PruneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Prune Subtree Endpoint Api Games  Game Id  Prune Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prune Subtree Endpoint",
        "tags": [
          "games"
        ]
      }
    },
    "/api/games/{game_id}/regenerate-node": {
      "post": {
        "description": "Regenerate the current node by pruning it and re-advending from parent.",
        "operationId": "regenerate_node_api_games__game_id__regenerate_node_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegenerateNodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Node",
        "tags": [
          "games"
        ]
      }
    },
...
  }
}